Design and Detail view complete
It took me forever to figure out why my TextViews weren't wrapping
correctly, but it turns out I had the layout_width defined as
wrap_content, so that was overriding the gravity I had set and extending
the TextView further than it should. I figured out if I set that
layout_width to 0dp it will respect the boundary of gravity when
wrapping the content.